updated login API to return user id
