+++ spring-boot-backend/src/main/java/com/example/springbootbackend/controller/AccountController.java
@@ -3,7 +3,6 @@ package com.example.springbootbackend.controller;
 import com.example.springbootbackend.auth.TokenService;
 import com.example.springbootbackend.dto.RequestErrorDTO;
 import com.example.springbootbackend.dto.account.AccountResponseDTO;
-import com.example.springbootbackend.dto.account.AccountLoginDTO;
 import com.example.springbootbackend.dto.account.AccountRequestDTO;
 import com.example.springbootbackend.service.account.AccountService;
 import lombok.extern.slf4j.Slf4j;
@@ -41,23 +40,6 @@ public class AccountController {
     public AccountResponseDTO getAccountById(@PathVariable Integer id) {
         log.info("Handling GET /api/accounts/{} request", id);
         return accountService.getAccountById(id);
-    }
-
-    // User account sign up/registration endpoint
-    @PostMapping("/signup")
-    public ResponseEntity<?> createAccount(@RequestBody AccountRequestDTO accountRequestDTO) {
-        log.info("Handling POST /api/accounts/signup request");
-        return new ResponseEntity<>(accountService.createAccount(accountRequestDTO), HttpStatus.CREATED);
-    }
-
-    // User account login endpoint
-    @PostMapping("/login")
-    public ResponseEntity<?> loginAccount(@RequestBody AccountLoginDTO accountLoginDTO) {
-        log.info("Handling POST /api/accounts/login request");
-        String token = accountService.loginAccount(accountLoginDTO);
-        Map<String, String> response = new HashMap<>();
-        response.put("token", token);
-        return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
     @PutMapping("/{id}")
